add new fragments logic
